Friendlier templates.
Added a get_input method to templates to get and store user input. If
you ask for the same thing twice, it will not ask the user twice, but
use the first input given.

Updated gitconfig to use get_input method.

Updated bash_profile to load all files from a ~/.bash.d directory. Make
this directory when running the install.

Added a gitignore dotfile and updated gitconfig to use this as a global
ignore file.