Bash-it
Added Bash-it for more bash goodness. Completions, aliases, and plugins
FTW! Using modified bash_it.sh that doesn't load any of the custom stuff
or backwards comatibility with older versions of bash-it.

Updated install script to source instead of running in a different
shell. Needed to do this to work with the new bash-it install script.