Move brew bundle into rcup post hook
I realized I didn't need the dotfile command as this functionality could
be replaced by the post hook provided by rcm. Added a dotfile alias to
rcup.

Need to figure out if I want to try to auto-update the dotfiles repo on
up or not. I don't know how often this would change would be useful.