Enable Homebrew packages that I use often
Replace dns-flush with an alias that uses the m-cli command. This is
aware of what commands are needed for different MacOS versions.

Turn Bluetooth off as part of the setup. I don't usually use it. Add
aliases for off/on in case I do.

Attempt restart at end of setup script.

Update hostname for personal laptop.