Move over "plugins" to bash.d directory
This is an old pattern I was following, where I would stick things I
wanted to load into bash per environment in ~/.bash.d. RCM actually
doesn't symlink the directory, but each file, so if someone adds a file
to ~/.bash.d, it doesn't automatically get added to the repo.