Use dtrx in replace of extract function
dtrx is a little more complete than the extract function.
Alias extract to dtrx, because I won't remember that.

Because the extract file was deleted, I found an issue when trying to
load broken symlinks. To fix this, I'm checking before I source it and
also removing broken symlinks in post-up.