disable docker prompt info
There's a current issue where it always displays. It doesn't add a ton
of value while it's broken, so disable it until it's fixed

Add a check to see if rcup is running before auto-updating. When my
machine reboots, it's likely that multiple shells will be open and it
caused issues last time it happened